+++ routes/branding.js
@@ -1,0 +1,19 @@
+var express = require('express');
+var router = express.Router();
+var mongoose = require('mongoose');
+
+
+const Branding = mongoose.model('Branding');
+
+
+
+
+
+
+
+
+
+
+
+
+module.exports = router;
