+++ routes/consumption.js
@@ -1,0 +1,26 @@
+var express = require('express');
+var router = express.Router();
+var mongoose = require('mongoose');
+
+
+
+const Consumption = mongoose.model('Consumption');
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+module.exports = router;
