+++ routes/events.js
@@ -1,0 +1,25 @@
+var express = require('express');
+var router = express.Router();
+var mongoose = require('mongoose');
+
+
+const Events = mongoose.model('Events');
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+module.exports = router;
