+++ routes/payments.js
@@ -1,0 +1,24 @@
+var express = require('express');
+var router = express.Router();
+var mongoose = require('mongoose');
+
+
+
+const Payments = mongoose.model('Payments');
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+module.exports = router;
