+++ routes/services.js
@@ -1,0 +1,24 @@
+var express = require('express');
+var router = express.Router();
+var mongoose = require('mongoose');
+
+
+
+const Services = mongoose.model('Services');
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+module.exports = router;
